+++ esig/__init__.py
@@ -26,14 +26,7 @@ except AttributeError:
     pass
 
 from esig.backends import get_backend, set_backend, list_backends
-
-try:
-    from esig.tosig import recombine
-    NO_RECOMBINE = False
-except ImportError:
-    NO_RECOMBINE = True
-    def recombine(*args, **kwargs):
-        raise NotImplementedError
+from esig.tosig import recombine
 
 
 __all__ = [
