+++ esig/backends.py
@@ -8,7 +8,7 @@ import threading
 import numpy
 
 # try:
-from . import _tosig
+from . import _tosig as tosig
 # except ImportError:
 #    # Error occurs during build sequence, since tosig does not exist
     # tosig = None
@@ -17,8 +17,6 @@ from . import _tosig
 try:
     import iisignature
 except ImportError:
-    if tosig is None:
-        raise ImportError("No available backend for signature calculations, please reinstall esig.")
     iisignature = None
